requirements.txt
flask
flask-cors
PyPDF2
google-generativeai
